requirements.txt
streamlit>=1.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.3.1
streamlit-sortables>=0.3.0
streamlit-calendar>=1.2.0
pyyaml>=6.0
bcrypt>=4.0.0
python-docx>=1.1.0
